+++ src/components/MessageList.js
@@ -5,8 +5,8 @@ import SVG from 'react-inlinesvg';
 
 import { shortenEthAddr } from '../utils';
 
-import Message from './Messages';
 import LoadingAnimation from './LoadingAnimation';
+import Message from './Messages';
 import closeIcon from '../assets/close-icon-black.svg';
 
 class MessageList extends Component {
@@ -21,10 +21,10 @@ class MessageList extends Component {
       currentUserAddr,
       colorTheme,
       userProfileURL,
-      isJoiningThread,
       isShowOnlineList,
       membersOnline,
       handleShowOnlineList,
+      isJoiningThread,
     } = this.props;
 
     return (
@@ -74,6 +74,7 @@ class MessageList extends Component {
             })}
           </div>
         </div>
+
         <div className={`sc-message-list ${isJoiningThread ? 'isLoading' : ''} ${isShowOnlineList ? '' : 'show'}`} ref={el => this.scrollList = el}>
           {isJoiningThread && (
             <LoadingAnimation colorTheme={colorTheme} threadLoading={isJoiningThread} />
@@ -118,8 +119,8 @@ MessageList.propTypes = {
   colorTheme: PropTypes.string,
   userProfileURL: PropTypes.func,
   handleShowOnlineList: PropTypes.func.isRequired,
-  isJoiningThread: PropTypes.bool,
   isShowOnlineList: PropTypes.bool,
+  isJoiningThread: PropTypes.bool,
 };
 
 export default MessageList;