+++ RELEASE-NOTES.md
@@ -1,5 +1,10 @@
 # Release Notes
 
+## v0.0.5 - 2020-01-09
+* chore: update 3box lib to 1.16.1
+* feat: add loading animation on first fetch
+* feat: render messages without joining thread
+
 ## v0.0.4 - 2020-01-13
 * fix: move node-sass from devDep to dep